Removed comment depth, enforced depth limit of 64k, added soft limit on depth.

+++ libraries/chain/include/steemit/chain/comment_object.hpp
@@ -3,7 +3,7 @@
 #include <steemit/protocol/authority.hpp>
 #include <steemit/protocol/steem_operations.hpp>
 
-#include <steemit/chain//steem_object_types.hpp>
+#include <steemit/chain/steem_object_types.hpp>
 #include <steemit/chain/witness_objects.hpp>
 
 #include <boost/multi_index/composite_key.hpp>
@@ -127,7 +127,7 @@ namespace steemit {
             time_point_sec active; ///< the last time this post was "touched" by voting or reply
             time_point_sec last_payout;
 
-            uint8_t depth = 0; ///< used to track max nested depth
+            uint16_t depth = 0; ///< used to track max nested depth
             uint32_t children = 0; ///< used to track the total number of children, grandchildren, etc...
 
             /**
